implemented WeatherFragment for ViewPager
